Fix zlib issue with the last reads of save files
For some reason zlib always gives Z_DATA_ERROR on the last call to
gzread for a savefile. The last 4 bytes do seemingly not match the
checksums correctly. These bytes are simply skipped, because the reason
for it being so is unknown. If it is desired to create a save file
editor at some point, these bytes might need to be appended at the end.