WIP CmakeRC and creature portraits
These images were intended to be displayed while hovering a graph showing
the best creature for each player over time. This stopped when the bytes
for the gifs were read in and no decompression method for gifs were available
at hand. There are many alternatives if this route is to be picked up again
later (e.g. write your own).

This branch is left stale in a WIP state for now.