Move plot data to separate library h3plots
This is a huge restructure of the code. The current implementation had
both the plot data ("backend") and the drawing of the data ("frontend")
inside the h3viewer application. This intertwining of code is not good
if one ever wants to another frontend.

Thus all the code for the plot data is separated into a new static
library, h3plots, which can then be interfaced from any application
wanting to draw the data. The only current API is to use the
h3plots::Plots class and pass a h3plots::SaveFileSeries to it to
initialize the data.